Added CLASS and ARRAY map parsing
