+++ src/api/youtube/requests/multichannel.php
@@ -1,19 +1,21 @@
 <?php
 
-namespace yt\request;
+namespace yt\youtube\request;
 
 use yt\interfaces\requestInterface;
 use yt\quota;
-use yt\response;
+use yt\youtube\response;
 
-class yt_playlists implements requestInterface
+class multichannel implements requestInterface
 {
-    public $description = "Performs a search on youtube channel for all playlists.";
+    public $description = "Returns list of video uploads from each channel supplied.";
 
-    public $parameters = 'None';
+    public $parameters = '
+    (array) [ \'channels\' => \'channelid1,channelid2\']
+    ';
 
     public $cost = 1;
-    public $cost_per_result = 2;
+    public $cost_per_result = 4; // snippet + contentParts
 
     public $domain = 'https://www.googleapis.com/youtube/v3';
 
@@ -27,6 +29,11 @@ class yt_playlists implements requestInterface
 
     public $response;
 
+    public $channel_list;
+
+    public $last_response;
+
+
     public function config($config)
     {
         $this->config = $config;
@@ -35,7 +42,7 @@ class yt_playlists implements requestInterface
 
     public function response()
     {
-        return $this->response;
+        return $this->response[0];
     }
 
     public function get_cost()
@@ -46,30 +53,28 @@ class yt_playlists implements requestInterface
 
     public function request()
     {
-        $this->build_request_url();
+        $this->create_channel_list_from_csv();
 
         (new quota)->update_quota_by_api_key($this->cost, $this->config['api_key']);
 
-        (new \yt\e)->line('- Calling API.', 1);
+        foreach ($this->channel_list as $channel_id) {
+            (new \yt\e)->line('- Calling API for Channel_ID : '.$channel_id, 1);
+            $this->build_request_url($channel_id);
+            $this->call_api();
 
-        try {
-            $this->response = json_decode(wp_remote_fopen($this->built_request_url));
-        } catch (\Exception $e) {
-            (new \yt\e)->line('- \Exception calling YouTube' . $e->getMessage(), 1);
-            return false;
-        }
-        
-        (new \yt\r)->last('search', 'RESPONSE:'. json_encode($this->response, JSON_PRETTY_PRINT));
+            if (!(new response)->is_errored($this->last_response)) {
+                return false;
+            }
 
-        if (!(new response)->is_errored($this->response)) {
-            return false;
+            (new quota)->update_quota_by_api_key($this->cost_per_result, $this->config['api_key']);
         }
 
-        $this->cost_of_items();
+        $this->combine_results();
+
+        $this->filter_for_uploads_only();
 
         return true;
     }
-
 
 
     // ┌─────────────────────────────────────────────────────────────────────────┐
@@ -81,19 +86,65 @@ class yt_playlists implements requestInterface
     // └─────────────────────────────────────────────────────────────────────────┘░
     //  ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
 
-    private function build_request_url()
+
+    private function combine_results()
     {
-        if(!$this->check_url()){return false;}  
-        $this->built_request_url = $this->domain . '/playlists?' . $this->config['query_string'] . "&key=" . $this->config['api_key'];
+        foreach ($this->response as $key => $response) {
+            if ($key == 0) {
+                continue;
+            }
+            $this->response[0]->items = array_merge($this->response[0]->items, $response->items);
+            unset($this->response[$key]);
+        }
     }
 
 
-    private function cost_of_items()
+
+    public function filter_for_uploads_only()
     {
-        $item_count = count($this->response->items);
-        (new quota)->update_quota_by_api_key(($this->cost_per_result * $item_count), $this->config['api_key']);
+        foreach ($this->response[0]->items as $key => $item) {
+            if (!isset($item->contentDetails->upload)) {
+                unset($this->response[0]->items[$key]);
+            }
+        }
     }
 
+
+
+    private function call_api()
+    {
+        try {
+            $this->last_response = json_decode(wp_remote_fopen($this->built_request_url));
+            $this->response[] = $this->last_response;
+        } catch (\Exception $e) {
+            (new \yt\e)->line('- \Exception calling YouTube' . $e->getMessage(), 1);
+            return false;
+        }
+        (new \yt\r)->last('search', 'RESPONSE:'. json_encode($this->last_response, JSON_PRETTY_PRINT));
+    }
+
+
+
+    private function create_channel_list_from_csv()
+    {
+        str_replace(' ', '', $this->config['extra_parameters']['channels']);
+        $this->channel_list = explode(',', $this->config['extra_parameters']['channels']);
+        return;
+    }
+
+
+
+    private function build_request_url($channel_id)
+    {
+        if (!$this->check_url()) {
+            return false;
+        }
+        if (!$channel_id) {
+            return false;
+        }
+        $this->built_request_url = $this->domain . '/activities?part=snippet%2CcontentDetails&channelId='.$channel_id.'&' . $this->config['query_string'] . "&key=" . $this->config['api_key'];
+        (new \yt\r)->last('search', 'QUERSTRING = '. $this->built_request_url);
+    }
 
     // ┌─────────────────────────────────────────────────────────────────────────┐
     // │                                                                         │░
@@ -103,6 +154,7 @@ class yt_playlists implements requestInterface
     // │                                                                         │░
     // └─────────────────────────────────────────────────────────────────────────┘░
     //  ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
+
 
     public function check_url()
     {
