+++ src/api/youtube/requests/subscriptions.php
@@ -1,18 +1,19 @@
 <?php
 
-namespace yt\request;
+namespace yt\youtube\request;
 
 use yt\interfaces\requestInterface;
 use yt\quota;
-use yt\response;
+use yt\youtube\response;
 
-class yt_search implements requestInterface
+class subscriptions implements requestInterface
 {
-    public $description = "Performs a search on the youtube search:list endpoint.";
+    public $description = "Performs a search of subscriptions for a channel.";
 
-    public $parameters = 'none';
+    public $parameters = 'None';
 
-    public $cost = 100;
+    public $cost = 1;
+    public $cost_per_result = 2;
 
     public $domain = 'https://www.googleapis.com/youtube/v3';
 
@@ -47,6 +48,8 @@ class yt_search implements requestInterface
     {
         $this->build_request_url();
 
+        (new quota)->update_quota_by_api_key($this->cost, $this->config['api_key']);
+
         (new \yt\e)->line('- Calling API.', 1);
 
         try {
@@ -55,16 +58,14 @@ class yt_search implements requestInterface
             (new \yt\e)->line('- \Exception calling YouTube' . $e->getMessage(), 1);
             return false;
         }
-
-        (new \yt\r)->last('search', 'RESPONSE:'. json_encode($this->response, JSON_PRETTY_PRINT));
         
+        (new \yt\r)->last('search', 'RESPONSE:'. json_encode($this->response, JSON_PRETTY_PRINT));
+
         if (!(new response)->is_errored($this->response)) {
             return false;
         }
 
-
-        (new quota)->update_quota_by_api_key($this->cost, $this->config['api_key']);
-
+        $this->cost_of_items();
 
         return true;
     }
@@ -83,8 +84,14 @@ class yt_search implements requestInterface
     private function build_request_url()
     {
         if(!$this->check_url()){return false;}  
-        $this->built_request_url = $this->domain . '/search?' . $this->config['query_string'] . "&key=" . $this->config['api_key'];
-        (new \yt\r)->last('search', 'QUERSTRING = '. $this->built_request_url); 
+        $this->built_request_url = $this->domain . '/subscriptions?' . $this->config['query_string'] . "&key=" . $this->config['api_key'];
+    }
+
+
+    private function cost_of_items()
+    {
+        $item_count = count($this->response->items);
+        (new quota)->update_quota_by_api_key(($this->cost_per_result * $item_count), $this->config['api_key']);
     }
