+++ src/scraper.php
@@ -180,17 +180,12 @@ class scraper
         // Pass all of the available substitutions into
         // the class so we can do some swapping.
         $this->api->set_substitutions($this->options->substitutions);
-
-        // set the search string
-        // Scrape Instance -> search array -> search string
-        $this->api->set_query($this->options->scrape[$this->_scrape_key]['yt_scrape_search']['yt_search_string']);
-
-        // What type of request is this?
-        $this->api->set_request_type($this->options->scrape[$this->_scrape_key]['yt_scrape_search']['yt_search_type']);
-
-        // Add any extra parameters passed through.
-        $this->api->set_extra_parameters($this->options->scrape[$this->_scrape_key]['yt_scrape_search']['yt_search_parameters']);
         
+        // Instead of passing each thing individually, just pass the array of
+        // config details instead. This means that any new parameters will
+        // automatically be passed through.
+        $this->api->set_search_config($this->options->scrape[$this->_scrape_key]['yt_scrape_search']);
+
         // Get the YouTube results and add to scrape array.
         $this->options->scrape[$this->_scrape_key]['yt_scrape_response'] = $this->api->run();
 
@@ -218,6 +213,11 @@ class scraper
 
         // do some checks.
         if ($this->has_response() == false) {
+            return;
+        }
+        if ($this->options->scrape[$this->_scrape_key]['yt_scrape_filter']['yt_filter_id'] == 'none')
+        {
+            $this->options->scrape[$this->_scrape_key]['yt_scrape_filtered'] = $this->options->scrape[$this->_scrape_key]['yt_scrape_response'];
             return;
         }
 
@@ -268,6 +268,10 @@ class scraper
         if ($this->has_filtered() == false) {
             return;
         }
+        if ($this->options->scrape[$this->_scrape_key]['yt_scrape_mapper']['yt_mapper_id'] == 'none')
+        {
+            return;
+        }
 
         // Give the mapper all transforms
         // This is because we'll need the parameters for
@@ -313,6 +317,10 @@ class scraper
         $this->importer = new import;
 
         if ($this->has_response() == false) {
+            return;
+        }
+        if ($this->options->scrape[$this->_scrape_key]['yt_scrape_import'] == 'none')
+        {
             return;
         }
 
@@ -368,6 +376,11 @@ class scraper
         $this->scheduler = new scheduler;
 
         (new \yt\e)->line('[ Scheduler ] : '.$this->options->scrape[$this->_scrape_key]['yt_scrape_schedule']['yt_schedule_id']);
+
+        if ($this->options->scrape[$this->_scrape_key]['yt_scrape_schedule']['yt_schedule_id'] == 'none')
+        {
+            return;
+        }
 
         $this->scheduler->set_scrape_id($this->options->scrape[$this->_scrape_key]['yt_scrape_id']);
 
@@ -472,7 +485,7 @@ class scraper
 
     public function has_filtered()
     {
-        if ($this->options->scrape[$this->_scrape_key]['yt_scrape_filtered'] == null) {
+        if (!isset($this->options->scrape[$this->_scrape_key]['yt_scrape_filtered'])) {
             (new \yt\e)->line('- There is no filtered array results to map and import.', 1);
             return false;
         }
